// blacksmith-of-lenai-app/src/app/api-generated/model/playerGardenField.ts
/**
 * Garden Game API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GardenCrop } from './gardenCrop';


export interface PlayerGardenField { 
    playerGardenFieldId: string;
    fieldIndex: number;
    isPurchased: boolean;
    isIrrigated: boolean;
    isFertilized: boolean;
    price: number;
    timePassed: number;
    gardenCrop?: GardenCrop;
}