// blacksmith-of-lenai-app/src/app/api-generated/model/gardenCrop.ts
/**
 * Garden Game API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { JobType } from './jobType';


export interface GardenCrop { 
    gardenCropId: string;
    gardenCropName: string;
    duration: number;
    energyCostsPlant: number;
    energyCostsHarvest: number;
    energyCostsFertilize: number;
    energyCostsIrrigation: number;
    energyCostsReplant: number;
    jobType: JobType;
    experience: number;
    jobExperience: number;
}
export namespace GardenCrop {
}